% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caclulate_error.R
\name{mae}
\alias{mae}
\title{Calcuate Root Mean Absolute Error}
\usage{
mae(actual, predicted)
}
\description{
Function that returns Root Mean Absolute Error (MAE)
}
\details{
Details go here
}
\examples{
\dontrun{}


}

