% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validations.R
\docType{data}
\name{merged}
\alias{merged}
\title{Visualise downscaled data}
\format{An object of class \code{NULL} of length 0.}
\usage{
merged
}
\description{
Function that returns Root Mean Square Error (RMSE)
}
\details{
Details go here
}
\examples{
\dontrun{}


}
\keyword{datasets}

