% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{parse_data}
\alias{parse_data}
\title{Parse 2015/16 Meteorological Verification Data}
\usage{
parse_data()
}
\description{
This function automates parsing of temperature and precipitation for
Queensland weather stations.
}
\details{
Details about the resulting files go here

The CSV or ESRI format shapefile in the respective year-directory
will contain the following fields/values:
\describe{
\item{station_number}{Station number (WMO/DATSAV3 number) for the location}
\item{area_code}{}
\item{start}{}
\item{end}{}
\item{Temperature}{Degrees Celsius}
\item{Rainfall}{Millimetres}
\item{LAT}{Latitude}
\item{LON}{Longitude}
\item{STN_HT}{Station reported elevation (metres to hundredths)}
\item{Aviation ID}{}
\item{MSAS Elevation}{}
\item{Distance from GridPt}{}
\item{Roughness}{}
\item{Distance from coast}{}
\item{Category}{}
\item{Forecast district}{}
\item{sa_special}{}
}
}
\note{
Users of these data should take into account the following
}
\examples{
\dontrun{}


}
\references{
{}
}

