% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QLD_hourly_locations}
\alias{QLD_hourly_locations}
\title{QLD_hourly_locations}
\format{A data frame with 293 observations of 4 variables:
\describe{
\item{WMO_NUM}{}
\item{station_number}{}
\item{station_name}{}
\item{LATITUDE}{}
\item{LONGITUDE}{}
\item{STN_HT}{}
\item{AVIATION_ID}{}
\item{REGION}{}
\item{GridPt Lat}{}
\item{GridPt Lon}{}
\item{MSAS elevation}{}
\item{Distance from GridPt}{}
\item{Roughness}{}
\item{Distance from coast}{}
\item{Category}{}
\item{forecast_district}{}
\item{sa_special}{}
}}
\usage{
QLD_hourly_locations
}
\description{
QLD_hourly_locations
}
\keyword{datasets}
