% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QLD_SILO_and_hourly_stations}
\alias{QLD_SILO_and_hourly_stations}
\title{QLD_SILO_and_hourly_stations}
\format{A list of 366 data frame objects, each containing one day of hourly
weather observations from 2015-04-30 to 2016-04-29
\describe{
  \item{station_number}{}
  \item{area_code}{}
  \item{parameter}{}
  \item{valid_start}{}
  \item{valid_end}{}
  \item{value}{}
  \item{unit}{}
  \item{statistic}{}
  \item{instantaneous}{}
  \item{level}{}
  \item{qc_valid_minutes}{}
  \item{qc_valid_minutes_start}{}
  \item{qc_valid_mintues_end}{}
}}
\usage{
QLD_SILO_and_hourly_stations
}
\description{
QLD_SILO_and_hourly_stations
}
\keyword{datasets}
