% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SILO_array}
\alias{SILO_array}
\title{SILO_array}
\format{An array of Queensland Scientific Information for Land Owners (SILO)
daily weather data, which have corresponding Bureau of Meteorology (BoM)
hourly data used for validation
\describe{
  \item{Station}{Station number}
  \item{Date}{Date (YYYYMMDD}
  \item{Day}{Day of year}
  \item{Date2}{}
  \item{T.Max}{Maximum daily air temperature record}
  \item{Smx}{}
  \item{T.Min}{Minimum daily air temperature recorded}
  \item{Smn}{}
  \item{Rain}{}
  \item{Srn}{}
  \item{Evap}{}
  \item{Sev}{}
  \item{Radn}{}
  \item{Ssl}{}
  \item{VP}{}
  \item{Svp}{}
  \item{RHmaxT}{}
  \item{RHminT}{}
  \item{FAO56}{}
}}
\usage{
SILO_array
}
\description{
SILO_array
}
\keyword{datasets}
