% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subscribers}
\alias{subscribers}
\title{subscribers}
\format{A table of example e-mail addresses subscribed to the e-mail service
\describe{
  \item{email}{User e-mail address}
  \item{my_location_name}{Users' location}
  \item{mylat}{Latitude}
  \item{mylon}{Longitude}
}}
\usage{
subscribers
}
\description{
subscribers
}
\keyword{datasets}
