% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validations.R
\name{plot_density}
\alias{plot_density}
\title{Visualise fit of downscaled data}
\usage{
plot_density(BoM_SILO_data)
}
\arguments{
\item{BoM_SILO_data}{Data object generated from using \code{downscale} to
generate hourly downscaled weather data from daily SILO data along with BoM
hourly data for validation}
}
\description{
Function that creates a hexbin density graph of observed versus predicted
hourly values.
}
\examples{
\dontrun{plot_density(BoM_SILO_data)}

}
