% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caclulate_error.R
\name{rmse}
\alias{rmse}
\title{Calcuate Root Mean Square Error}
\usage{
rmse(actual, predicted)
}
\description{
Function that returns Root Mean Square Error (RMSE)
}
\details{
Details go here
}
\examples{
\dontrun{}


}

