% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caclulate_error.R
\name{rmse}
\alias{rmse}
\title{Calculate Root Mean Square Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{observed values}

\item{predicted}{predicted values}
}
\description{
Function that calculates Root Mean Square Error (RMSE)
}
\details{
Details go here
}
\examples{
actual <- c(4, 6, 9, 10, 4, 6, 4, 7, 8, 7)
predicted <- c(5, 6, 8, 10, 4, 8, 4, 9, 8, 9)
rmse(actual, predicted)

}
